Improvements
Lots of little fixes, many are case specific. For example, if a city is
added to the my locations list, it becomes selected on that list, etc.
:coffee:
Also fixed Time class so that it will be correct with regard to daylight
savings, and updates to the current time upon refreshing the weather
data.

+++ src/main/java/Forecast24Hour.java
@@ -74,10 +74,10 @@ public class Forecast24Hour {
 			sclabel[i].setFont(minorfont);
 
 			if (preferences.getTempUnit().equalsIgnoreCase("F")){
-				templabel[i] = new JLabel(tempArray[i] + "°F");
+				templabel[i] = new JLabel((int)tempArray[i] + "°F");
 				templabel[i].setFont(minorfont);
 			} else {
-				templabel[i] = new JLabel(tempArray[i] + "°C");
+				templabel[i] = new JLabel((int)tempArray[i] + "°C");
 				templabel[i].setFont(minorfont);
 			}
